+++ src/mx_bluesky/common/external_interaction/ispyb/ispyb_utils.py
@@ -6,6 +6,7 @@ import os
 from ispyb import NoResult
 from ispyb.connector.mysqlsp.main import ISPyBMySQLSPConnector as Connector
 from ispyb.sp.core import Core
+
 from mx_bluesky.common.parameters.constants import SimConstants
