+++ tests/unit_tests/hyperion/external_interaction/conftest.py
@@ -1,63 +1,16 @@
 import os
-from collections.abc import Callable, Sequence
+from collections.abc import Sequence
 from copy import deepcopy
-from typing import Any
 from unittest.mock import MagicMock, mock_open, patch
 
-import bluesky.plan_stubs as bps
-import bluesky.preprocessors as bpp
 import pytest
-from bluesky.run_engine import RunEngine
 from ispyb.sp.mxacquisition import MXAcquisition
-from ophyd.sim import SynAxis
 
-from mx_bluesky.common.external_interaction.callbacks.plan_reactive_callback import (
-    PlanReactiveCallback,
-)
 from mx_bluesky.common.utils.utils import convert_angstrom_to_eV
 from mx_bluesky.hyperion.parameters.gridscan import HyperionThreeDGridScan
 from mx_bluesky.hyperion.parameters.rotation import RotationScan
 
 from ....conftest import raw_params_from_file
-
-
-class MockReactiveCallback(PlanReactiveCallback):
-    activity_gated_start: MagicMock
-    activity_gated_descriptor: MagicMock
-    activity_gated_event: MagicMock
-    activity_gated_stop: MagicMock
-
-    def __init__(self, *, emit: Callable[..., Any] | None = None) -> None:
-        super().__init__(MagicMock(), emit=emit)
-        self.activity_gated_start = MagicMock(name="activity_gated_start")  # type: ignore
-        self.activity_gated_descriptor = MagicMock(name="activity_gated_descriptor")  # type: ignore
-        self.activity_gated_event = MagicMock(name="activity_gated_event")  # type: ignore
-        self.activity_gated_stop = MagicMock(name="activity_gated_stop")  # type: ignore
-
-
-@pytest.fixture
-def mocked_test_callback():
-    t = MockReactiveCallback()
-    return t
-
-
-@pytest.fixture
-def RE_with_mock_callback(mocked_test_callback):
-    RE = RunEngine()
-    RE.subscribe(mocked_test_callback)
-    yield RE, mocked_test_callback
-
-
-def get_test_plan(callback_name):
-    s = SynAxis(name="fake_signal")
-
-    @bpp.run_decorator(md={"activate_callbacks": [callback_name]})
-    def test_plan():
-        yield from bps.create()
-        yield from bps.read(s)  # type: ignore # See: https://github.com/bluesky/bluesky/issues/1809
-        yield from bps.save()
-
-    return test_plan, s
 
 
 @pytest.fixture
